ZUGFeRD — augment shipment.spec.js with ZUGFeRD invoice-PDF case; drop standalone spec
The standalone einvoice-zugferd.spec.js reimplemented the O2C flow and failed CI
("Missing fields M_Warehouse_ID, C_BPartner_ID, M_PricingSystem_ID, PaymentRule,
C_Currency_ID") because it got mandatory order fields wrong.

Fix: add the ZUGFeRD case directly into shipment.spec.js, reusing the proven
SalesOrderPage / ShipmentSchedulePage / InvoiceCandidatePage / InvoicePage
page-object flow (the only new code is the createMasterdata seller/buyer recipe
and the archived-PDF fetch + PdfValidator.validateZugferdAttachment assertion).

The masterdata mirrors the cucumber eInvoiceZugferdEmail.feature Background:
orgSeller (org-bpartner VATaxID + DE address + IBAN + default contact),
ZUGFeRD buyer (isEInvoiceRecipeint + eInvoiceType=Z + eInvoiceBuyerReference),
adProcessFlags IsPdfA3Output=true, mock report service sysconfig.

TEST-COVERAGE.md: add sub-section 4b for the ZUGFeRD variant; remove old section 43.
